Fix MSVS compiler warnings.
Use explicit casts and appropriate types for pointer arithmetic.

+++ nan.h
@@ -1083,7 +1083,7 @@ template <typename TypeName> static size_t _nan_base64_decode(
   const TypeName* srcEnd = src + srcLen;
 
   while (src < srcEnd && dst < dstEnd) {
-    int remaining = srcEnd - src;
+    ptrdiff_t remaining = srcEnd - src;
     char a, b, c, d;
 
     while (_nan_unbase64(*src) < 0 && src < srcEnd) src++, remaining--;
@@ -1215,7 +1215,7 @@ static NAN_INLINE(void* NanRawString(
       }
       NanSetPointerSafe<size_t>(
           datalen
-        , toStr->WriteAscii(to, 0, sz_ + term_len, flags));
+        , toStr->WriteAscii(to, 0, int(sz_ + term_len), flags));
       return to;
 #endif
     case Nan::BINARY:
@@ -1232,7 +1232,7 @@ static NAN_INLINE(void* NanRawString(
       {
         uint16_t* twobytebuf = new uint16_t[sz_ + term_len];
 
-        size_t len = toStr->Write(twobytebuf, 0, sz_ + term_len, flags);
+        size_t len = toStr->Write(twobytebuf, 0, int(sz_ + term_len), flags);
 
         for (size_t i = 0; i < sz_ + term_len && i < len + term_len; i++) {
           unsigned char *b = reinterpret_cast<unsigned char*>(&twobytebuf[i]);
@@ -1250,7 +1250,7 @@ static NAN_INLINE(void* NanRawString(
         toStr->WriteOneByte(
             reinterpret_cast<uint8_t *>(to)
           , 0
-          , sz_ + term_len
+          , int(sz_ + term_len)
           , flags));
       return to;
 #endif
@@ -1263,7 +1263,7 @@ static NAN_INLINE(void* NanRawString(
       }
       NanSetPointerSafe<size_t>(
           datalen
-        , toStr->WriteUtf8(to, sz_ + term_len, NULL, flags) - term_len);
+        , toStr->WriteUtf8(to, int(sz_ + term_len), NULL, flags) - term_len);
       return to;
     case Nan::BASE64:
       {
@@ -1294,7 +1294,7 @@ static NAN_INLINE(void* NanRawString(
         int bc = 2 * toStr->Write(
             reinterpret_cast<uint16_t *>(to)
           , 0
-          , sz_ + term_len
+          , int(sz_ + term_len)
           , flags);
         NanSetPointerSafe<size_t>(datalen, bc);
         return to;
